Reuse runner pane
Start the runner pane with `remain-on-exit` option, so that
when an error is thrown during the Evaluate phase,
the error is clearly indicated in the runner pane.
We also respawn the runner pane when further edits are
made in the editor pane. This allows the new output
(or a new error) to be displayed in the runner pane.

'hsandbox' is terminated when the editor pane is terminated.